+++ heylogs-api/src/main/java/nbbrd/heylogs/Config.java
@@ -17,5 +17,8 @@ public class Config {
     String versioningId;
 
     @Nullable
+    String versioningArg;
+
+    @Nullable
     String forgeId;
 }