+++ heylogs-ext-calver/src/main/java/internal/heylogs/ext/calver/CalVerToken.java
@@ -1,0 +1,4 @@
+package internal.heylogs.ext.calver;
+
+public interface CalVerToken {
+}